move BLAS utils into BLAS module
