+++ base/deprecated.jl
@@ -1254,6 +1254,8 @@ end
 
 @deprecate showcompact_lim show
 
+@deprecate blas_set_num_threads BLAS.set_num_threads
+
 # During the 0.5 development cycle, do not add any deprecations below this line
 # To be deprecated in 0.6
