+++ base/exports.jl
@@ -620,7 +620,6 @@ export
 # linear algebra
     bkfact!,
     bkfact,
-    blas_set_num_threads,
     blkdiag,
     chol,
     cholfact!,
