+++ test/runtests.jl
@@ -15,7 +15,7 @@ cd(dirname(@__FILE__)) do
     if net_on
         n = min(8, Sys.CPU_CORES, length(tests))
         n > 1 && addprocs(n; exeflags=`--check-bounds=yes --depwarn=error`)
-        blas_set_num_threads(1)
+        BLAS.set_num_threads(1)
     end
 
     @everywhere include("testdefs.jl")
